+++ packages/core/src/interfaces/config.ts
@@ -1,5 +1,6 @@
 import { DocgeniLibrary } from './library';
 import { Locale } from './locale';
+import { DocItemToc } from './navigation-item';
 
 export interface HeroAction {
     text: string;
@@ -91,6 +92,8 @@ export interface DocgeniConfig {
     footer?: string;
     /** algolia config */
     algolia?: DocgeniAlgoliaConfig;
+    /** default toc, default is content **/
+    toc?: DocItemToc;
 }
 
 // For Angular Template