+++ packages/core/src/types/meta.ts
@@ -1,4 +1,4 @@
-import { HomeDocMeta } from '../interfaces';
+import { DocItemToc, HeadingLink, HomeDocMeta } from '../interfaces';
 
 export interface ComponentDocMeta {
     title?: string;
@@ -11,6 +11,7 @@ export interface ComponentDocMeta {
     label?: string;
     lastUpdatedTime?: number;
     contributors?: string[];
+    toc?: DocItemToc;
 }
 
 export interface CategoryDocMeta {