+++ packages/template/src/built-in/loader.ts
@@ -2,4 +2,6 @@ import { setBuiltInComponents } from './built-in-components';
 import label from './label/label.component';
 import alert from './alert/alert.component';
 
-setBuiltInComponents([label, alert]);
+export function loadBuiltInComponents() {
+    setBuiltInComponents([label, alert]);
+}