+++ packages/template/src/services/global-context.ts
@@ -3,6 +3,7 @@ import { DocgeniSiteConfig, NavigationItem, DocgeniMode, HomeDocMeta } from '../
 import { HttpClient } from '@angular/common/http';
 import docsearch from 'docsearch.js';
 import { languageCompare } from '../utils/language-compare';
+import { DOCUMENT } from '@angular/common';
 export const CONFIG_TOKEN = new InjectionToken('DOC_SITE_CONFIG');
 
 export const DEFAULT_CONFIG: DocgeniSiteConfig = {
@@ -23,8 +24,8 @@ export class GlobalContext {
 
     docItems: NavigationItem[];
     homeMeta: HomeDocMeta;
-    readonly owner: string;
-    readonly repo: string;
+    owner: string;
+    repo: string;
     get isDefaultLocale() {
         return this.locale === this.config.defaultLocale;
     }
@@ -33,20 +34,35 @@ export class GlobalContext {
         return !!(this.config.algolia && this.config.algolia.apiKey && this.config.algolia.indexName);
     }
 
-    constructor(@Inject(CONFIG_TOKEN) public config: DocgeniSiteConfig, private http: HttpClient) {
-        this.locale = config.defaultLocale;
-        const maybeLocale = window.localStorage.getItem(DOCGENI_LOCALE_KEY) || window.navigator.language || '';
-        const cacheMode = window.localStorage.getItem(DOCGENI_MODE_KEY);
-        if (maybeLocale) {
-            const isSupport = (config.locales || []).findIndex(item => languageCompare(item.key, maybeLocale));
+    constructor(@Inject(CONFIG_TOKEN) public config: DocgeniSiteConfig, private http: HttpClient, @Inject(DOCUMENT) private document: any) {
+        this.setup();
+    }
 
-            if (isSupport !== -1) {
-                this.locale = config.locales[isSupport].key;
+    private getLocaleKey() {
+        const localeKeyFromUrl = this.getLocalKeyFromUrl();
+        if (localeKeyFromUrl) {
+            return localeKeyFromUrl;
+        } else {
+            const cacheLocale = window.localStorage.getItem(DOCGENI_LOCALE_KEY) || window.navigator.language || '';
+            const locale = (this.config.locales || []).find(locale => {
+                return languageCompare(locale.key, cacheLocale);
+            });
+            if (locale) {
+                return locale.key;
+            } else {
+                return this.config.defaultLocale;
             }
         }
+    }
+
+    private setup() {
+        this.setLocale(this.getLocaleKey());
+
+        const cacheMode = window.localStorage.getItem(DOCGENI_MODE_KEY);
         if (cacheMode && ['lite', 'full'].includes(cacheMode)) {
-            config.mode = cacheMode as DocgeniMode;
+            this.config.mode = cacheMode as DocgeniMode;
         }
+
         document.body.classList.add(`dg-mode-${this.config.mode}`, `dg-theme-${this.config.theme}`);
         const pattern = /https:\/\/github.com\/([^\/]*)\/([^\/]*)/.exec(this.config.repoUrl);
         if (pattern && pattern.length === 3) {
@@ -55,7 +71,14 @@ export class GlobalContext {
         }
     }
 
-    setLocale(locale: string) {
+    public getLocalKeyFromUrl() {
+        const localeFromUrl = (this.config.locales || []).find(locale => {
+            return this.document.location.pathname.startsWith(`/${locale.key}`);
+        });
+        return localeFromUrl && localeFromUrl.key;
+    }
+
+    public setLocale(locale: string) {
         this.locale = locale;
         window.localStorage.setItem(DOCGENI_LOCALE_KEY, locale);
     }
@@ -70,7 +93,7 @@ export class GlobalContext {
                 next: (response: { navs: NavigationItem[]; docs: NavigationItem[]; homeMeta: HomeDocMeta }) => {
                     this.homeMeta = response.homeMeta;
                     this.navs = response.navs;
-                    this.docItems = this.flatNavs(this.navs);
+                    this.docItems = this.sortDocItems(this.navs);
                     resolve(response);
                 },
                 error: error => {
@@ -84,12 +107,16 @@ export class GlobalContext {
         return path.startsWith('/') ? `assets/content${path}` : `assets/content/${path}`;
     }
 
-    flatNavs(navs: NavigationItem[]) {
+    sortDocItems(navs: NavigationItem[]) {
         navs = navs.slice();
-        const list = [];
+        const list: NavigationItem[] = [];
         while (navs.length) {
             const item = navs.shift();
             if (item.items) {
+                item.items.forEach(child => {
+                    child.ancestors = child.ancestors || [];
+                    child.ancestors.push(...(item.ancestors || []), item);
+                });
                 navs.unshift(...item.items);
             } else if (!item.hidden) {
                 list.push(item);