+++ packages/template/src/shared/doc-meta/doc-meta.component.ts
@@ -1,22 +1,26 @@
 import { HttpClient } from '@angular/common/http';
-import { Component, Input, OnInit, HostBinding } from '@angular/core';
+import { Component, Input, HostBinding, OnChanges } from '@angular/core';
 import { DocItem } from '../../interfaces';
 import { GlobalContext } from '../../services/global-context';
 import { filter } from 'rxjs/operators';
 
 @Component({
     selector: 'dg-doc-meta',
-    templateUrl: './doc-meta.component.html'
+    templateUrl: './doc-meta.component.html',
+    host: {
+        class: 'dg-doc-meta'
+    }
 })
-export class DocMetaComponent implements OnInit {
-    @HostBinding(`class.dg-doc-meta`) isDocContribution = true;
+export class DocMetaComponent implements OnChanges {
+    @HostBinding(`class.dg-d-none`) hideDocMeta = true;
     @Input() docItem: DocItem;
     lastUpdatedTime: Date;
     contributors: string[];
+
     constructor(private http: HttpClient, private globalContext: GlobalContext) {}
 
-    ngOnInit() {
-        if (this.globalContext.owner && this.globalContext.repo) {
+    ngOnChanges(): void {
+        if (this.docItem.originPath && this.globalContext.owner && this.globalContext.repo) {
             this.http
                 .get(`https://api.github.com/repos/${this.globalContext.owner}/${this.globalContext.repo}/commits`, {
                     params: { path: this.docItem.originPath }
@@ -25,7 +29,12 @@ export class DocMetaComponent implements OnInit {
                 .subscribe((result: { author: { avatar_url: string; login: string }; commit: { author: { date: string } } }[]) => {
                     this.contributors = Array.from(new Set(result.map(item => item.author.login)));
                     this.lastUpdatedTime = new Date(result[0].commit.author.date);
+                    this.hideDocMeta = false;
                 });
+        } else {
+            this.contributors = undefined;
+            this.lastUpdatedTime = undefined;
+            this.hideDocMeta = true;
         }
     }
 }